Add more flexible keyword argument support
Function _kw_handler allows for a more flexible use with keyword
arguments. This approach was used to update both plotting functions as
well as the flare method of the Frequency class.
The information about the default values was added to the docstrings
which also were updated to have the same format as the other docstrings.