Separate Fingerprint and Heatmap, add recoloring
Again, the plotting tools are separated to emphasize that they are
more generally applicable.

The recolor method was added which allows to have multiple colormaps in
one matrix plot.
recolor expects a generator that yields the masking. The mask_gen
function is such a generator, but any generator could be used.

For now, no docstrings have been written yet.