Add highlight functionality to matrixplots
A highlight functionality is added. For this the labels of the seaborn
plots are split on '-' and placed in individual labels that are moved
based on the max extend of the labels.

The utility functions have been moved to the utils.py file such as the
max_extend, the mask_gen and the _kw_handler.
Still, no docstrings have been written, they will be moved from the
freq.py files soonish.

The plotting classes have been renamed and plotting functions have been
added that wrap the classes functionality and return an object of the
respective class.