Update and move loading & selecting functionality
The loading and selecting mechanic is still in the freq file. The class
is unnecessary which is why the main functionality has been updated and
placed into functions.

The read_tsv is exactly the same as in the freq file. The select
function on the other hand offers more functionality and is faster now.
Not only 2 ranges at max can be supplied now, but as many ranges as
wanted. Also the exact chosen numbers can be selected now instead of
just ranges. Both functionalities now are located in the utils file.

In the future the Frequency object most likely won't be maintained
anymore and then removed. For now, some plotting code still uses it so
it is not deleted yet. Warnings will be added to notify the users that
newer functions are available.