Fix marking and sorting with alphanumerical labels
The datastructure to communicate between selection strategies and
marking method is changed to a dictionary which contains a namedtuple
with assigned sorting number and tick location. The labeltext is the
key.

If the labels have numbers, the concatinated numbers are chosen for
sorting, else they are just enumerated.

The markings can be removed. Further they are resetted when new markings
are plotted. If more markings are desired, the marking dictionary can
just be extended. The legends are then also removed.

The range strategy is more intuitive, the range no longer has to be the
lowest/highest exact value in the data but the next higher / lower key
is automatically chosen. The function closest in utils achieves this.
The other two strategies now only assign the label if its not already
asigned. This way no duplicate labels are cluttering the legend.