Allow the repeated recoloring of matrixplots
To recolor, the values of the mesh is mapped to the new colormaps, and
the original array is overwritten so the new colors are plotted. This
led to errors upon multiple calls to the recolor_matrix method.

By saving the initial array the first time the function is called and
then later accessing that saved value, the recolor method now can be called
multiple times.