Fix aspect of dendrogram, fix alignment of labels
The aspect ratio is no longer set to equal but the figure size is
changed, in order to have equal aspect and not mess up the dendrogram
lines.

The number of labels on the left now is correctly computed for the
alignment. Now either 1 label, if all labels on the left hand side are
equal, or all labels are displayed.