Add easier flare labeling and change iteritems
Labels should be now more customizable in the ordering. Before hand it
always assumed to be sequence numbers if the labels include numbers.

Iteritems is deprecated in pandas after 1.5, so to be compatible with
more recent pandas versions change them to items instead.

Some hard to read code lines were replaced for easier to read list
comprehensions.